[backend] chore: move platform token util functions and new tests for validator
