+++ apps/backend/src/security/directive-graphql/validator/platform-token-validator.ts
@@ -1,81 +1,29 @@
 import express from 'express';
 import { v4 as uuidv4 } from 'uuid';
-import {
-  OrganizationCapability,
-  ServiceConfigurationStatus,
-} from '../../../__generated__/resolvers-types';
+import { OrganizationCapability } from '../../../__generated__/resolvers-types';
 import { requestContext } from '../../../context/request.context';
 import { PortalContext } from '../../../model/portal-context';
 import { UserLoadUserBy } from '../../../model/user';
-import { DeploymentRequestDomain } from '../../../modules/deployment/deployment.domain';
 import { OrganizationDomain } from '../../../modules/organization-management/organization/organization.domain';
 import { ServiceConfigurationDomain } from '../../../modules/registration/service-configuration/service-configuration.domain';
+import {
+  extractPlatformId,
+  extractPlatformToken,
+  validateAndGetRequestedPlatformToken,
+  validateExistsToken,
+} from '../../../modules/security-management/token/platform-token.util';
 import { PLATFORM_USER_EMAIL, PLATFORM_USER_UUID } from '../../../portal.const';
-import { logApp } from '../../../utils/app-logger.util';
 
-export const PLATFORM_TOKEN_HEADER = 'xtm-hub-platform-token';
-export const PLATFORM_ID_HEADER = 'xtm-hub-platform-id';
+export {
+  extractPlatformId,
+  extractPlatformToken,
+  PLATFORM_ID_HEADER,
+  PLATFORM_TOKEN_HEADER,
+  validateActivePlatformToken,
+  validateAndGetRequestedPlatformToken,
+  validateExistsToken,
+} from '../../../modules/security-management/token/platform-token.util';
 export const PLATFORM_TOKEN_DIRECTIVE_NAME = 'platform_token';
-
-export const extractPlatformToken = (req: express.Request): string | null => {
-  return (req?.headers?.[PLATFORM_TOKEN_HEADER] as string) || null;
-};
-
-export const extractPlatformId = (req: express.Request): string | null => {
-  return (req?.headers?.[PLATFORM_ID_HEADER] as string) || null;
-};
-
-const validateExistsToken = (req: express.Request): boolean => {
-  const token = extractPlatformToken(req);
-  if (!token) {
-    logApp.warn('[validatePlatformToken] Missing platformToken header');
-    return false;
-  }
-
-  const platformId = extractPlatformId(req);
-  if (!platformId) {
-    logApp.warn('[validatePlatformToken] Missing platformId header');
-    return false;
-  }
-  return true;
-};
-
-export const validateActivePlatformToken = async (
-  req: express.Request
-): Promise<boolean> => {
-  if (!validateExistsToken(req)) return false;
-
-  const serviceConfiguration =
-    await ServiceConfigurationDomain.loadConfigurationByPlatformAndToken({
-      platform_id: extractPlatformId(req),
-      token: extractPlatformToken(req),
-    });
-
-  return serviceConfiguration?.status === ServiceConfigurationStatus.Active;
-};
-
-export const validateAndGetRequestedPlatformToken = async (
-  req: express.Request
-) => {
-  if (!validateExistsToken(req)) return null;
-
-  const deploymentRequest =
-    await DeploymentRequestDomain.loadDeploymentRequestBy({
-      platform_token: extractPlatformToken(req),
-    });
-
-  const isValid =
-    deploymentRequest &&
-    deploymentRequest.platform_id !== null &&
-    deploymentRequest.platform_id === extractPlatformId(req);
-  if (!isValid) {
-    logApp.warn(
-      '[validatePlatformToken] No registration matching token, or invalid platformId provided'
-    );
-  }
-
-  return isValid ? deploymentRequest : null;
-};
 
 const loadOrganizationFromPlatformIdAndTokenHeaders = async (
   req: express.Request
